requirements.txt
Django>=3.2
speedtest-cli
gunicorn
